+++ src/Pim/Bundle/BaseConnectorBundle/Archiver/InvalidItemsCsvArchiver.php
@@ -77,4 +77,20 @@ class InvalidItemsCsvArchiver extends AbstractFilesystemArchiver
     {
         return 'invalid';
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supports(JobExecution $jobExecution)
+    {
+        if ($this->collector->getInvalidItems()) {
+            foreach ($this->collector->getInvalidItems() as $element) {
+                if (is_array($element)) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
 }